Added gpu-aware MPI version.
