+++ CMakeLists.txt
@@ -19,6 +19,7 @@ option(COSMA_WITH_BENCHMARKS "Generate the benchmark targets." ${MASTER_PROJECT}
 option(COSMA_WITH_INSTALL "Enable installation." ${MASTER_PROJECT})
 option(COSMA_WITH_PROFILING "Enable profiling." OFF)
 option(COSMA_WITH_NCCL "Use NCCL as communication backend." OFF)
+option(COSMA_WITH_GPU_AWARE_MPI "Use gpu-aware MPI for communication." OFF)
 option(BUILD_SHARED_LIBS "Build shared libraries." OFF)
 
 set(COSMA_BLAS "MKL" CACHE STRING
@@ -44,12 +45,15 @@ if (COSMA_WITH_NCCL AND NOT COSMA_BLAS IN_LIST GPU_BACKENDS)
     message(FATAL_ERROR "NCCL can only be used with the GPU backend.")
 endif()
 
+if (COSMA_WITH_GPU_AWARE_MPI AND NOT COSMA_BLAS IN_LIST GPU_BACKENDS) 
+    message(FATAL_ERROR "GPU-aware MPI can only be used with the GPU backend.")
+endif()
+
 # if using scalapack, then enforce building as a shared library
 # so that the function interpose can work
 #if (COSMA_SCALAPACK AND NOT BUILD_SHARED_LIBS)
 #    message(FATAL_ERROR "When SCALAPACK wrappers are used, COSMA has to be compiled as a shared library. Rerun cmake adding -DBUILD_SHARED_LIBS=ON.")
 # endif()
-
 
 # if SCALAPACK BLAS is not defined, then set it to be the same
 # as the scalapack provider
@@ -93,7 +97,7 @@ if (${COSMA_BLAS} STREQUAL "MKL" OR ${COSMA_BLAS} STREQUAL "CRAY_LIBSCI")
     endif()
 endif()
 
-project(cosma VERSION 2.5.1 LANGUAGES CXX C)
+project(cosma VERSION 2.6.0 LANGUAGES CXX C)
 
 include(GNUInstallDirs)
 
@@ -145,6 +149,10 @@ set(SCALAPACK_BLAS_TARGET "")
 
 set(NCCL_TARGET "")
 set(NCCL_DEF "")
+set(MPI_DEF "")
+
+set(GPU_TARGET "")
+set(GPU_DEFF "")
 
 # these are only GPU-backends
 if (${COSMA_BLAS} STREQUAL "CUDA" OR ${COSMA_BLAS} STREQUAL "ROCM")
@@ -158,21 +166,38 @@ if (${COSMA_BLAS} STREQUAL "CUDA" OR ${COSMA_BLAS} STREQUAL "ROCM")
     if (COSMA_WITH_NCCL)
         set(NCCL_DEF "COSMA_WITH_NCCL")
         if (${COSMA_BLAS} STREQUAL "CUDA")
-            find_package(CUDA)
+            find_package(CUDAToolkit)
             find_package(NCCL)
+            set(GPU_TARGET "CUDA::cudart")
             set(NCCL_TARGET "${NCCL_LIBRARIES}")
             message("NCCL INCLUDE DIRS = ${NCCL_INCLUDE_DIRS}")
             message("NCCL LIBRARIES = ${NCCL_LIBRARIES}")
             # set(NCCL_TARGET "${NCCL_LIBRARIES}")
         elseif (${COSMA_BLAS} STREQUAL "ROCM")
-            find_package(HIPLIBS)
+            find_package(hip)
             find_package(RCCL)
+            set(GPU_TARGET "hip::host")
+            set(GPU_DEF "-D__HIP_PLATFORM_HCC__")
             set(NCCL_TARGET "${RCCL_LIBRARIES}")
             message("RCCL INCLUDE DIRS = ${RCCL_INCLUDE_DIRS}")
             message("RCCL LIBRARIES = ${RCCL_LIBRARIES}")
             # set(NCCL_TARGET "${NCCL_LIBRARIES}")
         else()
             message(FATAL_ERROR "COSMA_WITH_NCCL specified, but no GPU backend chosen.")
+        endif()
+    endif()
+
+    if (COSMA_WITH_GPU_AWARE_MPI)
+        set(MPI_DEF "COSMA_WITH_GPU_AWARE_MPI")
+        if (${COSMA_BLAS} STREQUAL "CUDA")
+            find_package(CUDAToolkit)
+            set(GPU_TARGET "CUDA::cudart")
+        elseif (${COSMA_BLAS} STREQUAL "ROCM")
+            find_package(hip)
+            set(GPU_TARGET "hip::host")
+            set(GPU_DEF "-D__HIP_PLATFORM_HCC__")
+        else()
+            message(FATAL_ERROR "COSMA_WITH_GPU_AWARE_MPI specified, but no GPU backend chosen.")
         endif()
     endif()
 endif()
