+++ src/cosma/CMakeLists.txt
@@ -23,7 +23,10 @@ if(NOT TARGET cosma)
     if (GPU)
         list(APPEND cosma_src_files "pinned_buffers.cpp")
         if (COSMA_WITH_NCCL) 
-            list(APPEND cosma_src_files "gpu/utils.cpp")
+            list(APPEND cosma_src_files "gpu/nccl_utils.cpp")
+        endif()
+        if (COSMA_WITH_GPU_AWARE_MPI) 
+            list(APPEND cosma_src_files "gpu/gpu_aware_mpi_utils.cpp")
         endif()
     endif()
 
@@ -33,21 +36,32 @@ if(NOT TARGET cosma)
         $<BUILD_INTERFACE:${cosma_SOURCE_DIR}/src>
     )
     if (COSMA_WITH_NCCL)
-        target_include_directories(cosma PUBLIC ${NCCL_INCLUDE_DIRS})
+        target_include_directories(cosma PUBLIC ${NCCL_INCLUDE_DIRS}
+        )
     endif()
 
     target_compile_features(cosma PUBLIC cxx_std_14)
     target_link_libraries(cosma PUBLIC MPI::MPI_CXX
                                        costa
-                                PRIVATE ${BLAS_TARGET}
+                                       ${BLAS_TARGET}
     )
     if (COSMA_WITH_NCCL) 
-        target_link_libraries(cosma PUBLIC ${NCCL_TARGET})
+        target_link_libraries(cosma PUBLIC ${NCCL_TARGET}
+                                           ${GPU_TARGET}
+       )
     endif()
 
-    target_compile_definitions(cosma PRIVATE ${BLAS_DEF})
+    target_compile_definitions(cosma PUBLIC ${BLAS_DEF})
     if (COSMA_WITH_NCCL) 
-        target_compile_definitions(cosma PRIVATE ${NCCL_DEF})
+        target_compile_definitions(cosma PUBLIC ${NCCL_DEF}
+                                                 ${GPU_DEFF}
+        )
+    endif()
+    if (COSMA_WITH_GPU_AWARE_MPI) 
+        target_compile_definitions(cosma PUBLIC ${MPI_DEF}
+                                                ${GPU_DEFF}
+        )
+        target_link_libraries(cosma PUBLIC ${GPU_TARGET})
     endif()
 
     if(COSMA_WITH_PROFILING)
